Create three different files in lib folder
